+++ apps/ha-bff/graphql/types/mutation_type.rb
@@ -6,7 +6,8 @@ module Types
 
     field :onboard_patient,       mutation: Mutations::OnboardPatient
     field :update_patient,        mutation: Mutations::UpdatePatient
-    field :update_consent,        mutation: Mutations::UpdateConsent
+    field :update_consent,             mutation: Mutations::UpdateConsent
+    field :update_consent_directives,  mutation: Mutations::UpdateConsentDirectives
     field :schedule_appointment,  mutation: Mutations::ScheduleAppointment
     field :cancel_appointment,    mutation: Mutations::CancelAppointment
     field :create_encounter,      mutation: Mutations::CreateEncounter